--- requirements.txt
@@ -1,39 +1,0 @@
-alembic==0.9.6
-asn1crypto==0.24.0
-bcrypt==3.1.4
-beautifulsoup4==4.6.0
-blinker==1.4
-certifi==2017.11.5
-cffi==1.11.2
-chardet==3.0.4
-click==6.7
-cryptography==2.1.4
-Flask==0.12.2
-Flask-Login==0.4.1
-Flask-Migrate==2.1.1
-Flask-SQLAlchemy==2.3.2
-Flask-WhooshAlchemy==0.56
-Flask-WTF==0.14.2
-idna==2.6
-itsdangerous==0.24
-Jinja2==2.10
-lxml==4.1.1
-Mako==1.0.7
-MarkupSafe==1.0
-numpy==1.14.0
-pandas==0.22.0
-paramiko==2.4.0
-pyasn1==0.4.2
-pycparser==2.18
-PyNaCl==1.2.1
-python-dateutil==2.6.1
-python-editor==1.0.3
-pytz==2017.3
-requests==2.18.4
-six==1.11.0
-SQLAlchemy==1.2.0
-sshtunnel==0.1.2
-urllib3==1.22
-Werkzeug==0.13
-Whoosh==2.7.4
-WTForms==2.1
